requirements.txt
flask==3.0.0
azure-cosmos==4.5.1
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0  # For production deployment